% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quanteda.textplots-package.R
\docType{package}
\name{quanteda.textplots-package}
\alias{quanteda.textplots}
\alias{quanteda.textplots-package}
\title{quanteda.textplots: Plots for the Quantitative Analysis of Textual Data}
\description{
Plotting functions for visualising textual data. Extends 'quanteda' and related packages with plot methods designed specifically for text data, textual statistics, and models fit to textual data. Plot types include word clouds, lexical dispersion plots, scaling plots, network visualisations, and word 'keyness' plots.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/quanteda/quanteda.textplots/issues}
}

}
\author{
\strong{Maintainer}: Kenneth Benoit \email{kbenoit@lse.ac.uk} (\href{https://orcid.org/0000-0002-0797-564X}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Kohei Watanabe \email{watanabe.kohei@gmail.com} (\href{https://orcid.org/0000-0001-6519-5265}{ORCID})
  \item Haiyan Wang \email{whyinsa@yahoo.com} (\href{https://orcid.org/0000-0003-4992-4311}{ORCID})
  \item Adam Obeng \email{quanteda@binaryeagle.com} (\href{https://orcid.org/0000-0002-2906-4775}{ORCID})
  \item Stefan Müller \email{mullers@tcd.ie} (\href{https://orcid.org/0000-0002-6315-4125}{ORCID})
  \item Akitaka Matsuo \email{a.matsuo@lse.ac.uk} (\href{https://orcid.org/0000-0002-3323-6330}{ORCID})
}

Other contributors:
\itemize{
  \item Ian Fellows \email{ian@fellstat.com} (authored wordcloud C source code (modified)) [copyright holder]
  \item European Research Council (ERC-2011-StG 283794-QUANTESS) [funder]
}

}
\keyword{internal}
